+++ MockStock/src/main/java/io/gaboja9/mockstock/domain/payments/repository/PaymentHistoryRepository.java
@@ -1,5 +1,6 @@
-package io.gaboja9.mockstock.domain.payments;
+package io.gaboja9.mockstock.domain.payments.repository;
 
+import io.gaboja9.mockstock.domain.payments.entity.PaymentHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
